Full Angular Conversion

+++ src/app/pages/register/register.component.ts
@@ -1,6 +1,10 @@
 // @ts-nocheck  //Ignora erros TS (DEBUG)
 import { Component, OnInit } from '@angular/core';
-import Swal from 'sweetalert2';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
+import { Helpers } from 'src/app/shared/utils/herlpers';
+import { msg } from 'src/app/shared/utils/msg';
+import { RegisterService } from './services/register.service';
 
 @Component({
   selector: 'app-register',
@@ -9,7 +13,17 @@ import Swal from 'sweetalert2';
 })
 export class RegisterComponent implements OnInit {
 
-  constructor() { }
+  constructor(private fb: FormBuilder, private registerService: RegisterService, private router: Router) { }
+
+  msg = msg;
+  helpers = Helpers;
+  registerForm: FormGroup = this.fb.group({
+      role: ["", [Validators.required]],
+      fullname: ["", [Validators.required]],
+      birthdate: ["", [Validators.required]],
+      email: ["", [Validators.required, Validators.email]],
+      password: ["", [Validators.required]],
+  });
 
   ngOnInit(): void {
   }
@@ -24,55 +38,36 @@ export class RegisterComponent implements OnInit {
 }
 
 cadastrar() {
-    if (this.checkIfRoleIsChecked() == false) {
-        //alert("Selecione um perfil")
-        Swal.fire({
-            icon: 'error',
-            title: 'Oops...',
-            text: 'Pelo menos um dos perfis tem que ser selecionado',
-            footer: '<a href="">Why do I have this issue?</a>'
-        })
-        return;
-    }
-    // If here, the form is valid
-    // Grab form data
-    let payload = {
-        fullname: document.querySelector("#fullname").value,
-        birthdate: document.querySelector("#birthdate").value,
-        email: document.querySelector("#email").value,
-        password: document.querySelector("#password").value,
-        role: (document.getElementsByName("role")[0].checked ? "dev" : "cliente"), 
-    }
-
-    // Send Payload to API
-    fetch("https://626742f701dab900f1bcce82.mockapi.io/api/users", {
-        method: "POST",
-        body: JSON.stringify(payload),
-        headers: {
-            'Content-Type': 'application/json',
-        },
-    })
-    .then(response => response.json())
-    .then(response => {
-      //alert("Sucesso");
-        Swal.fire({
-            title: 'Sucesso',
-            text: 'Usuario Cadastrado',
-            icon: 'success',
-            confirmButtonText: 'Ok!'
-        }).then((result) => {
-            if(result.isConfirmed) {
-                // Redirect to project list
-                localStorage.setItem("user.name", response.fullname);
-                localStorage.setItem("user.id", response.id);
-                localStorage.setItem("user.role", response.role === "dev" ? "Desenvolvedor" : "Cliente");
-                window.location.href = "projects-list.html";
+    if(this.registerForm.valid) {
+        // Payload
+        let payload: IUser = this.registerForm.value;
+        this.registerService.postUser(payload).subscribe(
+            (response) => {
+                Swal.fire({
+                    title: 'Sucesso',
+                    text: 'Usuario Cadastrado',
+                    icon: 'success',
+                    confirmButtonText: 'Ok!'
+                }).then((result) => {
+                    if(result.isConfirmed) {
+                        // Redirect to project list
+                        localStorage.setItem("user.name", response.fullname);
+                        localStorage.setItem("user.id", response.id);
+                        localStorage.setItem("user.role", response.role === "dev" ? "Desenvolvedor" : "Cliente");
+                        
+                        this.router.navigateByUrl('list');
+                    }
+                })
             }
-        })
-    })
-    .catch(error => {
-        console.log(error);
-    });
+        );
+    }
+    else {
+        this.registerForm.markAllAsTouched();
+    }
+}
+
+toggleRole(role: 'dev' | 'cliente') {
+    this.registerForm.get('role')?.setValue(role);
 }
 
 }
